back/requirements.txt
fastapi>=0.128.0
fastmcp>=2.14.2
python-dotenv>=1.2.1
httpx>=0.28.1
aiosu>=2.3.5
pytest>=9.0.2
pytest-asyncio>=1.3.0
openai>=2.15.0
anthropic>=0.76.0
google-genai>=1.57.0